requirements.txt
tensorflow
pandas
numpy
scikit-learn
imblearn
netmiko
matplotlib
pyyaml